Feat: update edit function
